FIX : landing page link, auth flow
In this commit I have made the following changes . - Added proper redirects to the links of the landing page. - change the flex structure of landing page, not image is on the left side. - fixed auth flow, noe before redirecting, it checks if the location is not null and undefined. - tried to add glassmorphism to menu